+++ lib/rabl-rails.rb
@@ -42,6 +42,9 @@ module RablRails
   mattr_accessor :enable_jsonp_callbacks
   @@enable_jsonp_callbacks = false
 
+  mattr_accessor :allow_empty_format_in_template
+  @@allow_empty_format_in_template = false
+
   def self.configure
     yield self